+++ src/lexer/stream/stream.h
@@ -10,6 +10,6 @@ typedef struct
 extern const char* numbers;
 char CharacterStream_Next(CharacterStream_t *stream);
 CharacterStream_t CharacterStream_Create(const char *buffer);
-int CharacterStream_UntilNumEnd(CharacterStream_t* stream);
+long long CharacterStream_UntilNumEnd(CharacterStream_t* stream);
 void CharacterStream_UntilStrEnd(CharacterStream_t* stream, char* buffer);
 #endif
